Move to_fixed_hex and add test for string replacement
