+++ common/src/bytecode.rs
@@ -1,5 +1,4 @@
 use crate::errors::{BytecodeError, LinkError};
-use crate::str::AddressHexExt;
 use hex;
 use serde::de::{Error as DeError, Visitor};
 use serde::{Deserialize, Deserializer};
@@ -71,7 +70,7 @@ impl Bytecode {
         //   '__$name__..__' with the library address; see generated bytecode for
         //   `LinkedContract` contract for and example of how it looks like
         let placeholder = format!("__{:_<38}", name);
-        let address = address.to_fixed_hex();
+        let address = to_fixed_hex(&address);
         if self.0.find(&placeholder).is_none() {
             return Err(LinkError::NotFound(name.to_string()));
         }
@@ -194,6 +193,10 @@ impl<'de> Visitor<'de> for BytecodeVisitor {
     }
 }
 
+fn to_fixed_hex(address: &Address) -> String {
+    format!("{:040x}", address)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -206,5 +209,35 @@ mod tests {
     #[test]
     fn empty_hex_bytecode_is_empty() {
         assert!(Bytecode::from_hex_str("0x").unwrap().is_empty());
+    }
+
+    #[test]
+    fn to_fixed_hex() {
+        for (value, expected) in &[
+            (
+                "0x0000000000000000000000000000000000000000",
+                "0000000000000000000000000000000000000000",
+            ),
+            (
+                "0x0102030405060708091020304050607080900001",
+                "0102030405060708091020304050607080900001",
+            ),
+            (
+                "0x9fac3b52be975567103c4695a2835bba40076da1",
+                "9fac3b52be975567103c4695a2835bba40076da1",
+            ),
+        ] {
+            let value: Address = value[2..].parse().unwrap();
+            assert_eq!(super::to_fixed_hex(&value), *expected);
+        }
+    }
+
+    #[test]
+    fn string_replacement_replaces_multiple_occurrences() {
+        // This is important for Bytecode::link.
+        assert_eq!(
+            "ahellohellobhelloc".replace("hello", "world"),
+            "aworldworldbworldc"
+        );
     }
 }
